Fix and refactor higher-order multipole moments.
Previously there was a simple typo (Nibcint instead of Libcint) that made
the quadrupole to hexadecapole functions defunct.

I took that as an opportunity to refactor the multipole code.

I also added tests for the higher-order moments, based on reference data
generated with pyscf. This is not ideal, but I did not know where else to
easily get a reference.