shells: make shell size dynamic
Previously CartesianShell and SphericalShell used statically
sized SVectors. This leads to an uncanny mixture of type instability
– because in the basis they are stored in an inconcrete Vector{CartesianShell} –
and massive compilation costs whenever something decides to specialize on the shell size.

The latter is the case with Acsint. It basically needs to recompile entirely for
every invocation of `generate_ERI_quartet!` because of the many combinations of shell sizes.

This commit replaces the SVectors by Vectors, making Acsint and probably also
random other parts of GaussianBasis faster to compile and more type stable.